Finished Static portion, Close to edit button
