+++ src/src/App.js
@@ -1,0 +1,52 @@
+import Taskbar from "./Taskbar";
+import ListItems from "./ListItems";
+import BottomBar from "./BottomBar";
+import {useEffect, useState} from 'react';
+import React from "react";
+
+const initialData = [
+    {
+        id: 1,
+        folder: "Work",
+        tasks: [{
+            id: 4,
+            taskName: "Go To The Bank",
+            completed: false
+        }]
+    },
+    {
+        id: 2,
+        folder: "Personal",
+        tasks: [{
+            id: 5,
+            taskName: "Call Mom",
+            completed: false
+        }]
+    },
+    {
+        id: 3,
+        folder: "Academic",
+        tasks: [{
+            id: 6,
+            taskName: "Do CS 124",
+            completed: false
+        },
+            {
+                id: 7,
+                taskName: "Do CS 140",
+                completed: false
+            }
+        ]
+    }
+]
+
+function App() {
+    const [tasks, setEditTask] = useState(initialData)
+    return <div>
+        <Taskbar/>
+        <ListItems data={tasks} onTaskChanged={setEditTask}/>
+        <BottomBar/>
+    </div>
+}
+
+export default App;
