+++ src/src/BottomBar.js
@@ -1,7 +1,8 @@
 import './BottomBar.css';
+
 function BottomBar() {
     return <div id="bottom-bar">
-        <button id="add-btn" className="btn"><i className="fa-solid fa-plus"></i></button>
+        <button id="add-btn" className="btn">New Folder</button>
     </div>
 }
 export default BottomBar;
