+++ src/src/Folder.css
@@ -1,0 +1,11 @@
+/* Styling for the list items */
+#list-items {
+    color: black;
+    flex-basis: 80%;
+    width: 360px;
+    height: available;
+}
+
+ul {
+    padding-left: 0;
+}
