+++ src/src/ListItems.js
@@ -4,7 +4,7 @@ import Folder from './Folder';
 function ListItems(props) {
     return <div>
         {props.data.map(x =>
-            <Folder folder={x.folder} tasks={x.tasks}/>)}
+            <Folder folder={x.folder} tasks={x.tasks} onTaskChanged={props.onTaskChanged}/>)}
     </div>
 
 }
