+++ src/src/Task.css
@@ -1,0 +1,55 @@
+/* Positioning tasks */
+ul > ul {
+    padding-left: 20pt;
+}
+
+.tasks {
+    font-size: 20px;
+    background-color: #84CEEB;
+    color: black;
+    padding-left: 23px;
+    display: grid;
+    grid-template-columns: 1fr 30px 30px;
+}
+
+.tasks:nth-of-type(even) {
+    background-color: #5AB9EA;
+}
+
+/* line-through for completed taks */
+.completed{
+    text-decoration:line-through;
+    text-decoration-thickness: 2px;
+}
+
+/* Edit/Pencil and Paper button */
+.edit-btn {
+    border-radius: 12px;
+    grid-column: 2/span 1;
+    border: none;
+}
+
+.edit-btn:hover {
+    background-color: #666666;
+}
+
+/* Check mark button */
+.completed-btn {
+    border-radius: 12px;
+    grid-column: 3/span 1;
+    border: none;
+    font-weight: 100;
+}
+
+.completed-btn:hover {
+    background-color: #666666;
+}
+
+/* Background color of edit and check mark button */
+.even-btn {
+    background-color: #5AB9EA;
+}
+
+.odd-btn {
+    background-color: #84CEEB; ;
+}
